fix: Map field edge-case
During the check wheter a field is a "map", the string "entry" is added to
the field name, checked against the type name and then further checks are
made against the nested type of a parent message. In this edge-case, the
first check would pass even though it shouldn't and that would cause an
error because the parent type does not have a "nested_type" attribute.